Move audioOutput() implementations to impl.hpp-files
